refactor(step4) : 기능 검증 후 코드 정리
