--- src/main/java/ladder/domain/Ladder2.java
@@ -1,28 +1,0 @@
-package ladder.domain;
-
-import java.util.Collections;
-import java.util.List;
-
-public class Ladder2 {
-	private final List<LadderLine> ladder;
-
-	public Ladder2(List<LadderLine> ladder) {
-		this.ladder = ladder;
-	}
-
-	public int height() {
-		return ladder.size();
-	}
-
-	public List<LadderLine> getLadder() {
-		return Collections.unmodifiableList(ladder);
-	}
-
-	public Position resultByPlayer(Position PlayerPosition) {
-		Position position = PlayerPosition;
-		for (LadderLine ladderLine : ladder) {
-			position = ladderLine.move(position);
-		}
-		return position;
-	}
-}
