+++ src/main/java/nextstep/optional/Users.java
@@ -2,7 +2,6 @@ package nextstep.optional;
 
 import java.util.Arrays;
 import java.util.List;
-import java.util.Optional;
 
 public class Users {
     static final User DEFAULT_USER = new User("codesquad", 100);
