+++ src/test/java/ladder/domain/LadderResultTest.java
@@ -1,7 +1,6 @@
 package ladder.domain;
 
 import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
 
 public class LadderResultTest {
 	LadderGame ladderGame; // players, ladder
